moved jsonl file_io from recipe to clarity utils
