+++ recipes/cpc2/baseline/compute_haspi.py
@@ -13,22 +13,9 @@ from scipy.io import wavfile
 from tqdm import tqdm
 
 from clarity.evaluator.haspi import haspi_v2_be
+from clarity.utils.file_io import read_jsonl, write_jsonl
 
 logger = logging.getLogger(__name__)
-
-
-def read_jsonl(filename: str) -> list:
-    """Read a jsonl file into a list of dictionaries."""
-    with open(filename, "r", encoding="utf-8") as fp:
-        records = [json.loads(line) for line in fp]
-    return records
-
-
-def write_jsonl(filename: str, records: list) -> None:
-    """Write a list of dictionaries to a jsonl file."""
-    with open(filename, "a", encoding="utf-8") as fp:
-        for record in records:
-            fp.write(json.dumps(record) + "\n")
 
 
 def set_seed_with_string(seed_string: str) -> None:
