+++ tests/recipes/cpc2/test_compute_haspi.py
@@ -5,9 +5,7 @@ import pytest
 
 from recipes.cpc2.baseline.compute_haspi import (
     parse_cec2_signal_name,
-    read_jsonl,
     set_seed_with_string,
-    write_jsonl,
 )
 
 
@@ -42,34 +40,3 @@ def test_set_seed_with_string_ok(string_value):
     x = np.random.randint(0, 100)
     set_seed_with_string(string_value)
     assert np.random.randint(0, 100) == x
-
-
-@pytest.mark.parametrize(
-    "records",
-    [
-        [
-            {"a": 1, "b": 2},
-            {"a": 3, "b": 4},
-        ]
-    ],
-)
-def test_jsonl_write_read_loop(records, tmp_path):
-    """Test the write_jsonl and read_jsonl functions."""
-    write_jsonl(tmp_path / "test1.jsonl", records)
-    assert read_jsonl(tmp_path / "test1.jsonl") == records
-
-
-@pytest.mark.parametrize(
-    "records",
-    [
-        [
-            {"a": 1, "b": 2},
-            {"a": 3, "b": 4},
-        ]
-    ],
-)
-def test_jsonl_append_read_loop(records, tmp_path):
-    """Test the write_jsonl correctly appends to existing files."""
-    write_jsonl(tmp_path / "test2.jsonl", records)
-    write_jsonl(tmp_path / "test2.jsonl", records)
-    assert read_jsonl(tmp_path / "test2.jsonl") == records + records
